+++ yefer-garcia/tarea-5-puzzle/game.js
@@ -1,0 +1,103 @@
+import State from "./state.js";
+import Box from "./box.js";
+import  { swapBoxes,isSolved, getRandomGrid } from "./utils.js"
+
+class Game {
+    constructor(state) {
+      this.state = state;
+      this.tickId = null;
+      this.tick = this.tick.bind(this);
+      this.render();
+      this.handleClickBox = this.handleClickBox.bind(this);
+    }
+  
+    handleClickBox(box) {
+      return function () {
+        const nextdoorBoxes = box.getNextdoorBoxes();
+        const blankBox = nextdoorBoxes.find(
+          (nextdoorBox) => this.state.grid[nextdoorBox.y][nextdoorBox.x] === 0
+        );
+        if (blankBox) {
+          const newGrid = [...this.state.grid];
+          swapBoxes(newGrid, box, blankBox);
+          if (isSolved(newGrid)) {
+            clearInterval(this.tickId);
+            this.setState({
+              status: "won",
+              grid: newGrid,
+              move: this.state.move + 1,
+            });
+          } else {
+            this.setState({
+              grid: newGrid,
+              move: this.state.move + 1,
+            });
+          }
+        }
+      }.bind(this);
+    }
+  
+    tick() {
+      this.setState({ time: this.state.time + 1 });
+    }
+  
+    render() {
+      const { grid, move, time, status } = this.state;
+  
+      // Render grid
+      const newGrid = document.createElement("div");
+      newGrid.className = "grid";
+      for (let i = 0; i < 4; i++) {
+        for (let j = 0; j < 4; j++) {
+          const button = document.createElement("button");
+  
+          if (status === "playing") {
+            button.addEventListener("click", this.handleClickBox(new Box(j, i)));
+          }
+  
+          button.textContent = grid[i][j] === 0 ? "" : grid[i][j].toString();
+          newGrid.appendChild(button);
+        }
+      }
+      document.querySelector(".grid").replaceWith(newGrid);
+  
+      // Render button
+      const newButton = document.createElement("button");
+      if (status === "ready") newButton.textContent = "Play";
+      if (status === "playing") newButton.textContent = "Reset";
+      if (status === "won") newButton.textContent = "Play";
+      newButton.addEventListener("click", () => {
+        clearInterval(this.tickId);
+        this.tickId = setInterval(this.tick, 1000);
+        this.setState(State.start());
+      });
+      document.querySelector(".footer button").replaceWith(newButton);
+  
+      // Render move
+      document.getElementById("move").textContent = `Move: ${move}`;
+  
+      // Render time
+      document.getElementById("time").textContent = `Time: ${time}`;
+  
+      // Render message
+      if (status === "won") {
+        alert("Felicidades haz ganado!");
+      } else {
+        ("");
+      }
+    }
+  
+    setState(newState) {
+      this.state = { ...this.state, ...newState };
+      this.render();
+    }
+  
+    static ready() {
+      return new Game(State.ready());
+    }
+  
+    
+  }
+
+export default Game
+
