+++ yefer-garcia/tarea-5-puzzle/index.html
@@ -34,6 +34,7 @@
         <span id="time">Time: 100</span>
       </aside>
     </main>
-    <script src="main.js"></script>
+    <script src="main.js" type="module"
+    ></script>
   </body>
 </html>
