+++ yefer-garcia/tarea-5-puzzle/utils.js
@@ -1,0 +1,53 @@
+import Box from "./box.js";
+
+const swapBoxes = (grid, box1, box2) => {
+    const temp = grid[box1.y][box1.x];
+    grid[box1.y][box1.x] = grid[box2.y][box2.x];
+    grid[box2.y][box2.x] = temp;
+  };
+  
+const isSolved = (grid) => {
+    return (
+      grid[0][0] === 1 &&
+      grid[0][1] === 2 &&
+      grid[0][2] === 3 &&
+      grid[0][3] === 4 &&
+      grid[1][0] === 5 &&
+      grid[1][1] === 6 &&
+      grid[1][2] === 7 &&
+      grid[1][3] === 8 &&
+      grid[2][0] === 9 &&
+      grid[2][1] === 10 &&
+      grid[2][2] === 11 &&
+      grid[2][3] === 12 &&
+      grid[3][0] === 13 &&
+      grid[3][1] === 14 &&
+      grid[3][2] === 15 &&
+      grid[3][3] === 0
+    );
+  };
+  
+const getRandomGrid =()=> {
+    let grid = [
+      [1, 2, 3, 4],
+      [5, 6, 7, 8],
+      [9, 10, 11, 12],
+      [13, 14, 15, 0],
+    ];
+  
+    // Shuffle
+    let blankBox = new Box(3, 3);
+
+    grid.forEach(()=>{
+      const randomNextdoorBox = blankBox.getRandomNextdoorBox();
+      swapBoxes(grid, blankBox, randomNextdoorBox);
+      blankBox = randomNextdoorBox;
+    })
+    
+    if (isSolved(grid)) return getRandomGrid();
+    return grid;
+  };
+
+
+
+export  { swapBoxes, isSolved, getRandomGrid}
