+++ yefer-garcia/tarea2/style.css
@@ -36,8 +36,7 @@ a {
   flex-direction: column;
   justify-content: space-around;
 }
-#check {
-}
+
 .footer {
   align-items: center;
   background-color: rgb(31, 31, 31);
