+++ yefer-garcia/tarea3-globant/initializeMyName.js
@@ -1,11 +1,4 @@
 // initialize my name
 
-function initializeNames(name) {
-  var n = name.split(" ");
-  if (n.length < 3) return name;
-  for (i = 1; i < n.length - 1; i++) {
-    n[i] = n[i][0] + ".";
-  }
-  return n.join(" ");
-}
-console.log(initializeNames("Yefer Garcia "));
+const initializeNames = name => name.replace(/ (\w)\w*(?= )/g, ' $1.');
+initializeNames("Yefer Garcia  ")
